Fix outage detection an response + police lights on powerup
Increased shutdown timeout from 200 ms to 1.5s, which imitates the original fw more closely (200 ms was prone to false positives). Also, do not attempt to switch the relay off if it's already in an off state. Repeatedly doing so will drain the cap and cause voltage dropouts everywhere. The police lights things is useful for visually detecting reboots.